# backend/openapi.yaml
openapi: 3.0.0
paths:
...
  /api/v1/organizations/{organizationId}/projects/{tenantId}:
    get:
      operationId: ProjectsController_get
      parameters: []
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      tags:
        - projects
    patch:
      operationId: ProjectsController_update
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTenantDto'
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      tags:
        - projects
    delete:
      operationId: ProjectsController_remove
      parameters: []
      responses:
        '204':
          description: ''
      security:
        - bearer: []
      tags:
        - projects
...
  /api/v1/organizations/{organizationId}/ai-credentials/anthropic:
    get:
      operationId: AiCredentialsController_get
      parameters: []
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      tags:
        - ai credentials
    put:
      operationId: AiCredentialsController_set
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetAiCredentialDto'
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      tags:
        - ai credentials
    delete:
      operationId: AiCredentialsController_remove
      parameters: []
      responses:
        '204':
          description: ''
      security:
        - bearer: []
      tags:
        - ai credentials
  /api/v1/tenants/{tenantId}/environment-snapshots:
    post:
      operationId: CollectorEnvironmentSnapshotsController_receive
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnvironmentSnapshotV1Dto'
      responses:
        '202':
          description: ''
      security:
        - bearer: []
      tags:
        - collector
  /api/v1/organizations/{organizationId}/projects/{tenantId}/environment:
    get:
      operationId: EnvironmentSnapshotsController_status
      parameters: []
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      tags:
        - environment
  /api/v1/organizations/{organizationId}/projects/{tenantId}/environment-snapshots/{snapshotId}:
    get:
      operationId: EnvironmentSnapshotsController_get
      parameters: []
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      tags:
        - environment
  /api/v1/organizations/{organizationId}/projects/{tenantId}/analyses:
    get:
      operationId: AnalysesController_list
      parameters: []
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      tags:
        - analyses
  /api/v1/organizations/{organizationId}/projects/{tenantId}/analyses/{analysisId}:
    get:
      operationId: AnalysesController_get
      parameters: []
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      tags:
        - analyses
  /api/v1/organizations/{organizationId}/projects/{tenantId}/analyses/{analysisId}/work-items:
    get:
      operationId: AnalysesController_workItems
      parameters: []
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      tags:
        - analyses
  /api/v1/organizations/{organizationId}/projects/{tenantId}/analyses/{analysisId}/budget-approval:
    post:
      operationId: AnalysesController_approveBudget
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApproveBudgetDto'
      responses:
        '202':
          description: ''
      security:
        - bearer: []
      tags:
        - analyses
  /api/v1/organizations/{organizationId}/projects/{tenantId}/analyses/{analysisId}/resume:
    post:
      operationId: AnalysesController_resume
      parameters: []
      responses:
        '202':
          description: ''
      security:
        - bearer: []
      tags:
        - analyses
  /api/v1/organizations/{organizationId}/projects/{tenantId}/analysis-settings:
    get:
      operationId: AnalysesController_settings
      parameters: []
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      tags:
        - analyses
    put:
      operationId: AnalysesController_setSettings
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnalysisSettingsDto'
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      tags:
        - analyses
...
  /api/v1/organizations/{organizationId}/projects/{tenantId}/policies/{version}/edits:
    post:
      operationId: PolicyGenerationController_edit
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EditPolicyDto'
      responses:
        '202':
          description: ''
      security:
        - bearer: []
      tags:
        - policy generation
  /api/v1/organizations/{organizationId}/projects/{tenantId}/policy-generations:
    get:
      operationId: PolicyGenerationController_list
      parameters: []
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      tags:
        - policy generation
  /api/v1/organizations/{organizationId}/projects/{tenantId}/policy-generations/{attemptId}:
    get:
      operationId: PolicyGenerationController_get
      parameters: []
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      tags:
        - policy generation
...
info:
  title: Tessera Control Plane API
  description: Hosted control-plane API for Tessera
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    RoutingConfigurationDto:
      type: object
      properties:
        pathPrefix:
          type: string
          example: /
      required:
        - pathPrefix
    RuntimeThresholdsDto:
      type: object
      properties:
        requestTimeoutMs:
          type: number
          minimum: 100
          maximum: 120000
        maxRequestBodyBytes:
          type: number
          minimum: 0
          maximum: 104857600
      required:
        - requestTimeoutMs
        - maxRequestBodyBytes
    TenantRuntimeConfigurationDto:
      type: object
      properties:
        upstreamUrl:
          type: string
          example: https://app.internal.example
        failureBehavior:
          type: string
          enum:
            - allow
            - block
        unknownEndpointBehavior:
          type: string
          enum:
            - allow
            - block
        routing:
          $ref: '#/components/schemas/RoutingConfigurationDto'
        thresholds:
          $ref: '#/components/schemas/RuntimeThresholdsDto'
        samplingRate:
          type: number
          minimum: 0
          maximum: 1
      required:
        - upstreamUrl
        - failureBehavior
        - unknownEndpointBehavior
        - routing
        - thresholds
        - samplingRate
...
    UpdateTenantDto:
      type: object
      properties:
        name:
          type: string
          example: Checkout API
        slug:
          type: string
          example: checkout-api
        runtimeConfiguration:
          $ref: '#/components/schemas/TenantRuntimeConfigurationDto'
...
    SetAiCredentialDto:
      type: object
      properties:
        apiKey:
          type: string
          description: Anthropic API key. Write-only; never returned.
      required:
        - apiKey
    EnvironmentSnapshotV1Dto:
      type: object
      properties:
        schemaVersion:
          type: string
          enum:
            - tessera.environment/v1
    ApproveBudgetDto:
      type: object
      properties:
        ceilingUsd:
          type: number
          description: Spending ceiling in USD; the analysis stops when it is reached.
          example: 25
      required:
        - ceilingUsd
    AnalysisSettingsDto:
      type: object
      properties:
        autoApproveCeilingUsd:
          type: object
          nullable: true
          description: >-
            Analyses whose estimate fits under this ceiling start without manual
            approval. Null requires approval for every analysis.
      required:
        - autoApproveCeilingUsd
...
    EditPolicyDto:
      type: object
      properties:
        instruction:
          type: string
          maxLength: 2000
          description: >-
            Natural-language change to the base version. The result is a new
            version that must be reviewed; see precisionWarning.
      required:
        - instruction